requirements_live.txt
# Core dependencies for AiETHTrader AI Crypto Trading Pipeline
binance-connector>=3.5.1
pandas>=2.0.0
numpy>=1.24.0
pytz>=2023.3
torch>=2.2.0
scikit-learn>=1.3.0
joblib>=1.3.2
pyyaml>=6.0.1

# Additional useful packages
requests>=2.28.0
python-dotenv>=0.19.0
matplotlib>=3.5.0
seaborn>=0.11.0
python-dateutil>=2.8.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0

# Live trading specific
python-binance>=1.0.19
websocket-client>=1.6.0
